requirements.txt
phidata
python-dotenv
openai
duckduckgo-search
packaging
slack-sdk